move loading to common and add story
