+++ ui/src/components/common/Loading/Loading.stories.tsx
@@ -1,0 +1,9 @@
+import { Meta, StoryObj } from '@storybook/react';
+import { Loading } from './Loading';
+
+export default {
+  title: 'Loading',
+  component: Loading,
+} satisfies Meta;
+
+export const Default: StoryObj<typeof Loading> = {};
